refact: reorganize packages
