+++ app/src/main/java/br/com/mob1st/bet/features/competitions/domain/Competition.kt
@@ -1,4 +1,4 @@
-package br.com.mob1st.bet.features.competitions
+package br.com.mob1st.bet.features.competitions.domain
 
 import java.util.Date
