+++ app/src/main/java/br/com/mob1st/bet/features/competitions/domain/CompetitionRepository.kt
@@ -1,4 +1,4 @@
-package br.com.mob1st.bet.features.competitions
+package br.com.mob1st.bet.features.competitions.domain
 
 /**
  * Manage and execute CRUD operations in the [Competition] entity and everything that depends on
