+++ app/src/main/java/br/com/mob1st/bet/features/competitions/domain/Exceptions.kt
@@ -1,4 +1,4 @@
-package br.com.mob1st.bet.features.competitions
+package br.com.mob1st.bet.features.competitions.domain
 
 class GetDefaultCompetitionException(
     cause: Throwable
