+++ app/src/main/java/br/com/mob1st/bet/features/home/HomeScreen.kt
@@ -16,9 +16,9 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import br.com.mob1st.bet.features.competitions.CompetitionsTabScreen
+import br.com.mob1st.bet.features.competitions.presentation.CompetitionsTabScreen
 import br.com.mob1st.bet.features.groups.GroupsTabScreen
-import br.com.mob1st.bet.features.profile.ProfileTabScreen
+import br.com.mob1st.bet.features.profile.presentation.ProfileTabScreen
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
