+++ app/src/main/java/br/com/mob1st/bet/features/launch/LauncherViewModel.kt
@@ -6,7 +6,6 @@ import br.com.mob1st.bet.core.ui.state.AsyncState
 import br.com.mob1st.bet.core.ui.state.FetchedData
 import br.com.mob1st.bet.core.ui.state.SimpleMessage
 import br.com.mob1st.bet.core.ui.state.StateViewModel
-import kotlinx.coroutines.delay
 import org.koin.android.annotation.KoinViewModel
 
 @KoinViewModel
@@ -28,8 +27,6 @@ class LauncherViewModel(
     private fun triggerUseCase() {
         setState {
             launchAppUseCase()
-            // just as example. we have to replace it further
-            delay(1_500)
             it.data(data = LaunchData.finished.set(it.data, true))
         }
     }
