+++ app/src/main/java/br/com/mob1st/bet/features/profile/domain/User.kt
@@ -1,4 +1,4 @@
-package br.com.mob1st.bet.features.profile
+package br.com.mob1st.bet.features.profile.domain
 
 /**
  * The user.
